+++ src/functionalTest/resources/scenarios/RIA-7808_Judge_Decide_a_Costs_Home_Office.json
@@ -1,0 +1,90 @@
+{
+  "description": "RIA-7808: Home office lart has created the costings and a judge had  to costs which notifies legal rep",
+  "request": {
+    "uri": "/asylum/ccdAboutToSubmit",
+    "credentials": "Judge",
+    "input": {
+      "id": 7808,
+      "eventId": "decideCostsApplication",
+      "state": "appealSubmitted",
+      "caseData": {
+        "template": "minimal-appeal-submitted.json",
+        "replacements": {
+          "journeyType": "rep",
+          "appealType": "revocationOfProtection",
+          "legalRepresentativeEmailAddress": "{$TEST_LAW_FIRM_A_USERNAME}",
+          "legalRepReferenceNumber": "A1234567",
+          "appealGroundsRevocation": {
+            "values": [
+              "revocationRefugeeConvention",
+              "revocationHumanitarianProtection"
+            ]
+          },
+          "appliesForCosts": [
+            {
+              "id": "1",
+              "value": {
+                "appliedCostsType": "Unreasonable costs",
+                "respondentToCostsOrder": "Legal representative",
+                "applyForCostsHearingType": "No",
+                "scheduleOfCostsDocuments": [],
+                "applyForCostsCreationDate": "{$TODAY}",
+                "applyForCostsApplicantType": "Home office",
+                "argumentsAndEvidenceDetails": "",
+                "argumentsAndEvidenceDocuments": [
+                  {
+                    "id": "2f352b16-0ed9-4c2e-88c2-4bc091485504",
+                    "value": {
+                      "document_url": "http://dm-store:8080/documents/45ec3a85-8fad-4caa-81a5-6b3b5c3b3bb8",
+                      "document_filename": "test.pdf",
+                      "document_binary_url": "http://dm-store:8080/documents/45ec3a85-8fad-4caa-81a5-6b3b5c3b3bb8/binary"
+                    }
+                  }
+                ],
+                "applyForCostsHearingTypeExplanation": ""
+              }
+            }
+          ],
+          "legalRepName": "Fake Org Ltd",
+          "respondToCostsList": {
+            "value": {
+              "code": "11",
+              "label": "Costs 5, Wasted costs, 22 Nov 2023"
+            },
+            "list_items": [
+              {
+                "code": "15",
+                "label": "Costs 1, Wasted costs, 22 Nov 2023"
+              }
+            ]
+          }
+        }
+      }
+    }
+  },
+  "expectation": {
+    "status": 200,
+    "errors": [],
+    "caseData": {
+      "template": "minimal-appeal-submitted.json",
+      "replacements": {
+        "appealType": "revocationOfProtection",
+        "legalRepReferenceNumber": "A1234567",
+        "appealGroundsRevocation": {
+          "values": [
+            "revocationRefugeeConvention",
+            "revocationHumanitarianProtection"
+          ]
+        },
+        "legalRepName": "Fake Org Ltd"
+      }
+    },
+    "notifications": [
+      {
+        "reference": "7808__DECIDE_A_COSTS_EMAIL",
+        "recipient": "{$applyForCostsHomeOfficeEmailAddress}",
+        "subject": "Immigration and Asylum appeal: Costs application decided"
+      }
+    ]
+  }
+}
